fix:　名前の変更と処理の変更
